Move validator base class
